Fix installation flaw by making a proper dvcsautosync module
I installed dvcsautosync via sudo python setup.py install and it
installed everything into /usr/local/.... However the dvcs-autosync
script hardcoded a /usr/share/SCRIPT path as additional python path,
so the invokation failed and bailed out.

The solution is to make a proper dvcsautosync module that is
installed into the python path. This way we can call a local dvcs-autosync
or one in /usr/local or whereever and it will work as long as python can
find the dvcsautosync package. It also has the advantage that we don't
need to specify each single file in setup.py that needs to be installed.

I left the autostart script in /usr/local/share/dvcs-autosync, as I did
not know a better place for it.

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>