Deuglyify filenotifier code
Simply wrapping the whole code for the conditional platform
specific code in filenotifier was pretty ugly. Split all 3 OS
specific implementations into their own file and import only the
correct file. This reduces the indentation of the code by one and
makes things better to understand in my opinion (no need for me to
trawl through Windows specific code and to understand when which
code in that file will be available.

One issue though, dvcs-autosync failed when
pyinotify was not installed, but is that not optional? In that case,
where is the generic fallback code that we have? Or is this a hard
requirement now? Without pyinotify I get on master the error
"OS is not supported". On 'devel' I get simply a mysterious import
error that hides any mention of pyinotify. With this patch, you get
a nicer understandable error message, telling the user to install
python-inotify.

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>