Move pid file functions into helper functions
Create an Autosync class and put the create_and_check_pid_file and the
remove_pid_file helper functions there. This is nicer than having one
main huge function that does everything.

I know quite some ways in which the PID function can be simplified, but
that will be a separate step.

globals are evil! The secret plan is to continue moving stuff into the
Autosync class until everything is nicely encapsulated.

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>